# docs/openapi3.yaml
openapi: 3.0.3
info:
  title: Squad Admin Configurator
  version: 1.7.0
  description: Панель управления привилегированными пользователями в Squad
paths:
...
  /v1/api/privileged/servers/{id}/:
    get:
      operationId: v1_api_privileged_servers_retrieve
      description: View set для доступа к списку серверов
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this Сервер.
        required: true
      tags:
      - v1
      security:
      - tokenAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Server'
          description: ''
    put:
      operationId: v1_api_privileged_servers_update
      description: View set для доступа к списку серверов
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this Сервер.
        required: true
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Server'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Server'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Server'
        required: true
      security:
      - tokenAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Server'
          description: ''
    patch:
      operationId: v1_api_privileged_servers_partial_update
      description: View set для доступа к списку серверов
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this Сервер.
        required: true
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedServer'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedServer'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedServer'
      security:
      - tokenAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Server'
          description: ''
    delete:
      operationId: v1_api_privileged_servers_destroy
      description: View set для доступа к списку серверов
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this Сервер.
        required: true
      tags:
      - v1
      security:
      - tokenAuth: []
      responses:
        '204':
          description: No response body
...
  /v1/api/rotations/{url}/:
    get:
      operationId: v1_api_rotations_retrieve
      parameters:
      - in: path
        name: url
        schema:
          type: string
        required: true
      tags:
      - v1
      responses:
        '200':
          description: No response body
  /v1/api/rotations/{url}/{slug}/:
    get:
      operationId: v1_api_rotations_retrieve_2
      parameters:
      - in: path
        name: slug
        schema:
          type: string
        required: true
      - in: path
        name: url
        schema:
          type: string
        required: true
      tags:
      - v1
      responses:
        '200':
          description: No response body
components:
  schemas:
...
    PatchedServer:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        is_active:
          type: boolean
          title: Активирован
          description: Активирован ли сервер, если флаг не установлен - конфигурация
            этого сервера будет пустая
        title:
          type: string
          title: Название
          description: Название сервера
          maxLength: 50
        description:
          type: string
          title: Описание
          description: Описание сервера
          maxLength: 300
...
    Server:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        is_active:
          type: boolean
          title: Активирован
          description: Активирован ли сервер, если флаг не установлен - конфигурация
            этого сервера будет пустая
        title:
          type: string
          title: Название
          description: Название сервера
          maxLength: 50
        description:
          type: string
          title: Описание
          description: Описание сервера
          maxLength: 300
      required:
      - id
      - title
...
  securitySchemes:
    tokenAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Token-based authentication with required prefix "Token"